% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{zaro_list}
\alias{zaro_list}
\title{List contents of a Zarr store}
\usage{
zaro_list(store, path = "")
}
\arguments{
\item{store}{a store object from \code{\link[=zaro]{zaro()}}}

\item{path}{character. Path prefix within the store (default root).}
}
\value{
character vector of keys (relative paths).
}
\description{
List contents of a Zarr store
}
